Fix bugs
- Fix trust/untrust player NullPointerException
- Simplified math in InterestEventListener
- Fixed bug with negative offline bank loss displayed
- Fixed bugs in Database

+++ src/main/java/com/monst/bankingplugin/banking/account/Account.java
@@ -563,12 +563,16 @@ public class Account extends BankingEntity {
 
 	@Override
 	public void trustPlayer(OfflinePlayer p) {
+		if (p == null)
+			return;
 		super.trustPlayer(p);
 		plugin.getDatabase().addCoOwner(this, p, null);
 	}
 
 	@Override
 	public void untrustPlayer(OfflinePlayer p) {
+		if (p == null)
+			return;
 		super.untrustPlayer(p);
 		plugin.getDatabase().removeCoOwner(this, p, null);
 	}
